feat: translate the NotesList Activity into fragment
