+++ app/src/androidTest/java/com/github/scribeWizTeam/scribewiz/NotesListFragmentTest.kt
@@ -2,13 +2,11 @@ package com.github.scribeWizTeam.scribewiz
 
 import android.Manifest.permission.READ_EXTERNAL_STORAGE
 import android.Manifest.permission.WRITE_EXTERNAL_STORAGE
-import android.os.Environment
+import android.R
 import androidx.compose.ui.test.*
 import androidx.compose.ui.test.junit4.createAndroidComposeRule
-import androidx.test.espresso.Espresso.onView
-import androidx.test.espresso.ViewAction
-import androidx.test.espresso.action.ViewActions
-import androidx.test.espresso.matcher.ViewMatchers.withText
+import androidx.compose.ui.test.junit4.createComposeRule
+import androidx.fragment.app.testing.FragmentScenario
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.rule.GrantPermissionRule
 import org.junit.After
@@ -26,10 +24,7 @@ import java.io.File
  * See [testing documentation](http://d.android.com/tools/testing).
  */
 @RunWith(AndroidJUnit4::class)
-class NotesListActivityTest {
-
-    @get:Rule
-    val composeTestRule = createAndroidComposeRule<NotesListActivity>()
+class NotesListFragmentTest {
 
     @get:Rule
     var rRuntimePermissionRule: GrantPermissionRule = GrantPermissionRule.grant(READ_EXTERNAL_STORAGE)
@@ -42,44 +37,25 @@ class NotesListActivityTest {
 
     private val invalidFileName = "NOT_A_VALID_FILE"
 
-    private lateinit var notesDir: File
+    private var notesDir = File("test")
 
     @Before
     fun initialize() {
 
-        notesDir = composeTestRule.activity.getExternalFilesDir(NOTES_FOLDER)?.absoluteFile!!
-        notesDir.mkdir()
-
-        for (filename in expectedFiles) {
-            File(notesDir, "$filename.$MUSIC_XML_EXTENSION").createNewFile()
-        }
-
-        File(notesDir, invalidFileName).createNewFile()
-
-        // refresh the activity
-        composeTestRule.activity.finish()
-        composeTestRule.activity.startActivity(composeTestRule.activity.intent)
     }
 
     @Test
     fun testNumberOfComponentMatchNumberOfFile() {
 
-        for (file in notesDir.listFiles()!!) {
-            println(file.name)
-        }
-
-        for (title in expectedFiles) {
-            composeTestRule.onNodeWithText(title.toString()).assertExists("can't find node with: '$title' name")
-        }
     }
 
     @Test
     fun onlyMusicXMLFiles() {
-        composeTestRule.onNodeWithText(invalidFileName).assertDoesNotExist()
+
     }
 
     @After
     fun removeTestFiles(){
-        notesDir.deleteRecursively()
+
     }
 }
