+++ app/src/androidTest/java/com/github/scribeWizTeam/scribewiz/NotesStorageManagerTest.kt
@@ -1,13 +1,10 @@
 package com.github.scribeWizTeam.scribewiz
 
-import android.os.Environment
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import org.junit.After
 import org.junit.Assert.*
 import org.junit.Before
-import org.junit.Rule
 import org.junit.Test
-import org.junit.rules.TemporaryFolder
 import org.junit.runner.RunWith
 import java.io.File
 import kotlin.io.path.createTempDirectory
